net/sched: egdev_all, Register callbacks at the end of egdev_all list
Previously, newly registered callbacks were added at the head of the
egdev list, which resulted in the callbacks being called in the reverse
order the devices were registered in. Add them to the tail of the egdev
list instead, to maintain consistency between the register order and the
callback order.

Issue: 2324590
Change-Id: I14eead7611a10d2e9c96a1747b35c53cf4a6307f
Signed-off-by: Gavi Teitz <gavi@nvidia.com>